:tractor: relocate UknownTypeEncoder back to utils
